Add a patch version to the MaterialX versioning.
This change adds a patch version to augment the MaterialX version triplet.

Traditionally, MaterialX has had three items in the triplet. The Major and Minor are taken by the document model, whereas the build version has been the one incrementing for every release.

This has made it somewhat difficult to discern when a change is significant or minor. Recently, we've had two changes to MaterialX to rectify build issues, but these have also incremented the same number as a version doing a major update within the same document model version.

By adding a patch version, we allow MaterialX to better delineate to consumers when a change is significant or minor.

For this, I do not modify any areas that may affect symbols or namespaces, as the patches should effectively be transparent to those.

Signed-off-by: Dhruv Govil <dgovil2@apple.com>