Fix renaming of the ND_normalmap_vector2 node definition.
The upgrade code for the <normalmap> shader was indiscriminately
renaming all node definitions to "ND_normalmap_float". This included
"ND_normalmap_vecto2" nodes, whose nodedef should not be renamed.

This commit fixes that, limiting the renaming only to "ND_normalmap",
based on the type of the "scale" input, which is the only
differentiating factor between the two definitions.