Remove unit settings from distance nodedefs inputs (#1406)
Specifying both "unittype" and "unit" in nodedefs breaks something, and the conversion between units does not happen when switching them in Viewer.

This change remove "units" from nodedefs inputs of type "distance".

Note that "unit" remains for angles. Removing it causes an error (likely a bug) at load time.
"Unrecognized destination unit: meter"
Seems like materialX tries to convert degrees to meters.

While things seem to be still a bit buggy for angles, this seem to currently work at least for distences.
Viewer doesn not allow to change the angle unit, so it can't be tested in that environment.
It's likely that angle will fail as distances were failing when unit was present.