Marble and Speckle procedurals plus minor fixes (#1407)
Adding Marble and Speckle nodedefs and nodegraphs. This includes a small utility node called util_specklenoise. Note that this is classified as an utility node rather than a new noise because it's just an existing noise with some multiply and range adjustments that is re-used multiple times in Marble and Speckle. It has no real use outside of that.

Couple of more fixes
1) The legacy_noise procedural transform order has changed. Now rotate is execute before translate. (small node connection changes from line 440 to 510 in nodegraph file)

2) Small formatting/space changes to legacy_noise nodedef. (line 69 to 80)