Removing extra value attributes form nodgraphs inputs (#1405)
This fix was already done for the protein graphs, after an issue with LookdevX discovered it.

The MaterialX team made some changes to avoid double "value" and "interfacename" (or "nodename")  to the Graph Editor and added a sanity check on load.

Loading the Prism documents in 1.39 now throws a bunch of warnings. This fix removes all the extra "value" attributes and no warnings are generated any more.