Additional changes to the PreviewSurface Nodegraph
A recent change updated the nodegraph to reflect the updated PreviewSurface v2.5, which removes specular highlights for fully transparent materials. This change makes it such that opacity = 0 is treated the same regardless of if we use the transmissionOpacity or transmissionRefraction options, or if the material is in cutout or transparency mode.