Simplify the MetalTexture abstraction (#2095)
`mx_texture.metal` is used to provide an abstraction to the texture code when being called from GLSL code.

There are a few functions in the interface of this abstraction that do not appear to be used anywhere.

This PR proposes removing them, as a precursor to other upcoming texture work.  I'm proposing this as a separate PR to make things easier to review.

All of the MSL unit tests pass.