Named Constants and template support added to standard library Allowing for significant reduction in the size of the source for the standard library Templates can be nested and replace multiple variables simultaneously Templates can be expanded and diff in cmake stage - allowing for easy validation of appropriate template expansion Leaves the libraries folder untouched for now, to facilitate easier hands on experimentation with the new grammar
Also includes documentation building tools - which have been the proving ground for the new template grammar.
Builds specification docs to a temporary folder (for now) from a new template location - again for easy validation that
the expansion is correct.
The documentation work was critical to ensuring this initial template system is fully featured enough to
be useful, and uncovered a number of discrepancies with the human edited specification. This is one of the
strong advantages of using a template expansion based system.