Increase OSL samples in reference renders
This changelist increases the number of OSL samples in reference-quality renders, improving the visual parity between OSL and other languages in MaterialX render tests.

Additionally, it pares down the list of document paths for MaterialX render tests, reducing the resulting PDF document to a size that can be directly shared on Slack.