Update UsdPreviewSurface to version 2.6 (#2084)
This PR updates the PreviewSurface nodegraph to address the opacity issue in cutout mode described in gitIssue https://github.com/AcademySoftwareFoundation/MaterialX/issues/2044. It also updates the nodegraph to reflect the recently updated PreviewSurface v2.6 spec which adds an `opacityMode` attribute, to control the lighting response that fully transparent materials receive. Since the default behavior for this attribute is that fully transparent materials receive a lighting response, the PreviewSurface glass example remains unchanged.