Simplify HW shader generation for closures (#2205)
Currently all the GLSL BSDF functions are declared as separate functions per closure context (reflection, transmission, indirect, etc.).  This leads to some repetition in the GLSL code and some significant complexity in the HW Shader Generation code.

This PR refactors the HW Shader Generation to avoid the need for separate `ClosureContext` objects for each illumination stage, in favor of passing a single `ClosureData` object to each GLSL BSDF function, with a `closureType` enum indicating which illumination stage is being executed.