Replacement of Checker procedural (#1414)
The previous Checker was done in Graph Editor. This version imported Checker into LookdevX and re-authored it similarly to the other more recent procedurals.

The shader uses the stdlib Checker, which does not have a Rotation option. The result was that Rotation was processed before Translation, which is the opposite of OGS.
To address this, the stdlib Checker offset is not used any more, and the UV are translated and rotated before being sent to Checker.

Color outside of the pattern when not fully tiled is now black to match other procedurals.

Note that there are small inconsistencies in inputs and formats that will be addressed in bulk affecting multiple nodes at a later stage. Jiras already exist for that.