Remove references to CONFIGURE_DEPENDS
This changelist removes references to CONFIGURE_DEPENDS in our CMake configurations, as this feature is not well-implemented across our target platforms.

From the CMake documentation:

"The CONFIGURE_DEPENDS flag may not work reliably on all generators, or if a new generator is added in the future that cannot support it, projects using it will be stuck. Even if CONFIGURE_DEPENDS works reliably, there is still a cost to perform the check on every rebuild."