Adding Legacy Gradient procedural (#1416)
This is the protein Gradient, supporting all the gradient types and noise settings.
There is a base "float" version and a simple but usable "color 3" version that interpolates between 2 colors.
The Color3 version is not a Protein requirement, but the Ramp we need to match Protein is still TBD.
Something to deal with Ramp will be added later when updating to 1.39, and even then we will have a sub-par user experience to define a ramp properly. This is a discussion for another day.