Refine matrix multiplication tests to enable validation (#2308)
The `math_operators.mtlx` test for matrix multiplication was multiplying just identity matrices. If actually ran to produce an image, it would not show an error in matrices having swapped arguments.

In this PR I've changed the values so the result is `0.75, 0.25, 0.5, 1` when multiplies correctly, and `0.25, 0.75, 0.5, 1` when swapped.

I ran the tests with this `<input name="renderTestPaths" type="string" value="resources/Materials/TestSuite/stdlib/math" />`, before after the fix introduced in #2298.