Release all images in `MetalTextureHandler::releaseRenderResources` (#2316)
This PR addresses #2303.

This brings the MetalTextureHandler inline with the GLTextureHandler for releasing all the images when `nullptr` is passed.

Without this - both CPU and Metal texture memory is being leaked if the image caches are flushed.  

Recursively releasing the resources fixes the Metal texture memory leak, but that just highlighted the CPU side memory leak.  This is due to the MetalTextureHandler also retaining an `ImagePtr` in `_imageBindingInfo`.  Removing the `ImagePtr` from this storage fixes the memory leak, and it doesn't seem to be necessary to store the image, the only thing the image is used for is to access the resource ID - which is the same as the key to the `unordered_map`.