Fix Cloverleaf offset (#2327)
Cloverleaf Center was not doubled as other dimensions within the node, resulting in a different offset than other similar shape nodes (circle, hexagon).

The Tiledcloverleafs node, which uses Cloverleaf, is not affected by the fix as the scale and offsets of the tiled pattern is achieved through UVs and Center is always 0,0.