Swizzle removal (#1419)
Removing swizzle nodes by replacing the ones extracting "rgb" with a Convert colr4 to color 3, and the ones extracting "a" with an Extract using index 3.

The diff is confusing to read, but the changes are very simple.
Verified also by opening it back in LookdevX and Graph Editor, the connections are sound and the nodes types, inputs, and outputs are correct.