Adding missing units attributes (#1418)
Waves and Wood were missing the units attribute because I was not sure if they were supposed to be in units or not.
In CMUI (Inventor) they do not have units, but that seems odd as that would make them inconsistent if projects units were changed. Tests show that changing the units do not change the procedural size.
After experimenting a bit I'm sure units are needed even if not used in the current CMUI. For Waves, the units seem to be cm.
Also, tests shows that Waves is the only procedural working in World space. The default has been updated too.
More testing and refinements might be necessary once we can run legacy and graphs side by side.