Protein Procedurals Defs cleanup (#1420)
* Protein Procedurals Defs cleanup

After submitting each procedural separately, this update unifies the defs format.

Details:
- Added "_color3" suffix to 4 graphs that were missing it (and same changes in nodegraph file)
- Added all missing uinames, based mostly on CMUI naming
- Removed xpos/ypos from older defs created in Graph Editor
- All uifolder removed, we might use them one day, but for now they are just cluttering
- Moved attributes around to try to make them in the same or close to same order
- Added enum for Gradient Type (float and color3)
- Some unit="degree" removed. Works both ways, and we still have issues on that, but for now I unified all to no unit attribute for 3D angles
  (but stays for 2D rotation)
- All versions are set to "0.1" for now. We might decide for a real number one day.
- Utility nodes (the ones with "_util_" in teir name remains bare and that's ok, they are not meantto be used by users.
- Offsets XY and XYZ and rotation XYZ are excluded by changes since they will be replaced in the next round.

All these changes are cosmetic/UI, except for the 4 renames.
Next changes will be both on nodedefs and nodegraphs, once this is merged.

* Re-Adding degrees for Noise

Apparently, unit="degree" is still needed for Noise (and Noise only).
This is a bug in Materialx (active in the issues list) and I still don't know when the unit is needed not not.
Technically, it should not be there, but in this case it throws an error (he's trying to convert to "meters" for some reason).

* Removing version from stain node in Hardwood

Not sure why, but the node legacy_stain had a version when used in Hardwood.
This is the only instance I see in the whole NG file. Not sure what tool introduced it.
Removed and tested locally.