Add UDIM example and fix handling in Graph Editor (#2113)
Hoping to address #2007 - and also have some data to test while working on improving UDIM support in MaterialX.

Once I modified the textures it became clear that loading this in to MaterialXGraphEditor didn't actually load the correct images in to the correct materials.  They would all end up using the fallback material, so this PR includes a small fix to address that.