Improvements to closure addition in HW shading
This changelist proposes improvements to the computation of throughput in the GLSL implementation of closure addition, deriving the math from our implementation of albedo scaling in vertical layering.

I'm very interested in feedback and review on this proposal, in order to make sure the assumptions I'm making here are correct.