Improve node candidate filtering when adding a node from a link (#1542)
This changes improves the logic for filtering out the candidate nodes in the Add Node popup when the popup was triggered by drawing a link from a pin.

The first point of improvement is that the filtered candidate nodes are displayed now in the menu tree structure with categories rather than the linear list of candidate nodes.

The second point of improvement is that those if statements have simpler logic for and's and or's than the original code.