Filter out nodes having no inputs of matching type to the pin where the link was drawn from (#1542)
Adds a new member inputTypes, which is a set of input types of each NodeDef, to MenuItem of the GraphEditor. The inputTypes are used later to filter in/out candidate nodes in the Add Node menu popped up by drawing a link from a pin. The filtering is applied by matching the pin's type against the types of inputs of each candidate node.