Add auto link behavior to the new Add Node mechanism (#1542)
This change addes the ability to automatically link the newly created node back to the original pin when a node is created in the Add Node pop-up triggered by drawing a link from a pin.

As a part of that, the addLink() method has been split into checkCanAddLink() and addLink() to separate out the ed::AcceptNewItem() and ed::RejectNewItem() from the actual logic of link addition.