Fix Graph Editor crash with multioutput nodes (#2397)
When connecting multiple output nodes to surface shaders, `Graph::addLink` was not specifying the output being connected. As a consequence, the shader would end up using the first output on the upstream node which could cause shader compilation errors and exceptions if the input and output types did not match.

This change ensures that the output is specified on the input pin.