Unify `heighttonormal` node implementations
This changelist unifies the GLSL and OSL implementations of the `heighttonormal` node, using a simplified approach based on the gradient of the heightfield signal with respect to the input texture coordinates.

This approach makes the output normal map independent of screen-space size and orientation, and allows the `heighttonormal` node to be applied to procedural heightfields of arbitrary complexity, rather than restricting it to scalar images.

The specification for `heighttonormal` has been updated to clarify the role of the `texcoords` input and provide more guidance to renderer authors, though we should ultimately provide complete equations for the behavior of the node in our normative specification.