Allow keyboard deletion of selected links in Graph Editor (#2426)
This PR adds the ability to delete selected links using the keyboard, same as it works for nodes.
The existing functionality to Option-Click on a link is still the same, but now deletion of links should be more discoverable.

Personally I only found Option-click by looking at the code while trying to understand how I can delete links; the obvious select-and-press-delete hadn't worked, with this PR it does.