Protein Phase 3 - Ceramic Mirror Metal Glazing (#1425)
* Protein Phase 3 - Ceramic Mirror Metal Glazing

Revision of Ceramic, Mirror, Metal, Glazing Protein materials as part of the phase 3 of the graphs work.

Most changes are about input names, new inputs for normals, switch on enum inputs from float to integer (and associated NG changes), swizzle nodes removal, and other graphs optimizations and documentation (backdrops added).

These graphs are now authored in LookdevX, so a lot of changes in the diff are formatting and node coordinates.

The changes to the inputs will break the materials converter if this version of MaterialX goes into USD.

* Updating unit tests to reflect new defs

Updating input types and names for Ceramic, Metal, Glazing