Add deprecated `fps` input to `time` node
This changelist adds a deprecated `fps` input to the specification and implementations of the `time` node, providing continuity of node signatures for applications using earlier versions of the MaterialX data libraries.

In a future minor version upgrade, we can then formally remove this input, providing version upgrade logic to handle earlier assets.

In a quick survey of `time` node implementations, the `fps` input appears to be ignored, making this proposed approach a reasonable match with industry expectations:

https://developer.apple.com/documentation/shadergraph/application/time-(float)
https://www.sidefx.com/docs/houdini/nodes/vop/mtlxtime.html
https://github.com/mrdoob/three.js/pull/29149