differentiate between PDBs for shared vs. static libraries
On Windows, linked-generated PDB files are typically associated with
shared libraries and executables, and as a result would be installed
alongside them in the "bin" directory. Compiler-generated PDB files are
typically associated with static libraries and would be installed
alongside them in the "lib" directory.

See the CMake documentation for more detail on the distinction:
https://cmake.org/cmake/help/latest/prop_tgt/PDB_NAME.html