Add deprecated `fps` input to `time` node (#2450)
This changelist adds a deprecated `fps` input to the specification and implementations of the `time` node, providing continuity of node signatures for applications using earlier versions of the MaterialX data libraries.

In a future minor version upgrade, we can then formally remove this input, providing version upgrade logic to handle earlier assets.