Optimizations to `heighttonormal` (#2453)
This changelist implements optimizations to our gradient-based implementation of `heighttonormal` in GLSL and OSL, using a cross product in place of the inverse Jacobian.  The new version requires no division instructions at runtime, and generates virtually identical results.

Credit for this idea is due to @kwokcb, who originally suggested it in a [comment](https://github.com/AcademySoftwareFoundation/MaterialX/pull/2424#pullrequestreview-2893382794), and it was additionally inspired by recent discussions with Stephen Hill and Ron Radeztsky at the Lucasfilm ADG.