Optimization to mx_ggx_smith_G2
This changelist implements an optimization to the `mx_ggx_smith_G2` function in GLSL, replacing two divides with four multiplies.

Validation of this proposed optimization was conducted by rendering the `standard_surface_default.mtlx` example in the MaterialX Viewer on an NVIDIA RTX A6000 at 4K resolution, with the directional albedo mode set to Monte Carlo, and this change reduces the total frame time from 72ms to 70ms.