Treat checkbox-driven uniforms as bool in Viewer (#2437) (#2463)
This PR seeks to address #2437, wherein toggling a checkbox off and on in the Property Editor causes the corresponding uniform to stop updating, remaining in the "off" state indefinitely thereafter. The proposed fix is to pass the value of the checkbox to shaders as a bool rather than a float.

Tested in MaterialXView on macOS 15.5 (OpenGL and Metal rendering paths).