add additional stage to optimize combination of 3 bsdfs that have been mixed with 2 nodes.
Add possible refactor of gltf_pbr that allows it to match the existing optimization rule

Add checkbox to MaterialXView to control bsdf mix optimization

add GenOptions option to control the optimization

remove trailing underscores from node names - because these still make variables with double underscores.

Fix asan crash.

remove debug print

remove double underscores from node names - appears to be illegal.

Refactor OpenPBR nodegraph optimization in to a ShaderGenerator optimization step for ND_mix_bsdf.

Replace the mix with an add and modify the upstream weight inputs accordingly to perform the same math as a mix.  This allows the BSDF nodes to early out if their contribution is not necessary in the mix.