Remove Blur node implementation (#2411)
I'm submitting this PR - because I think it's the best we can do for now.  I looked in to trying to find a portable robust implementation for this, and couldn't really come up with anything.

I think it's better to be explicitly do nothing, than to do something that is confusing. 

MDL currently just passes through the `in` input as the output, so I'm using that as the template for the others as it seems like a reasonable fallback - implemented as a nodegraph that is now shared across all languages.