Fix dot node crash in code generation (#2505)
Code attempts to use `in` on `dot` not without checking for existence on the node.

It will not be exposed (added) as a uniform input if it's connected as is the case with the `dot` nodes
used in the open_pbr to standard_surface conversion graph.

Simply add in a firewall check for existence first.