Removing space input (#1440)
* Removing space input

The "space" input has been deprecated. To avoid issue we remove the input as the default value (tangent) is the desired one, so this will work with the current and the new nodedef.

* Removing space also form translation graphs

These graphs used "object" as the space input for normalmap.
After checking the correct input should have been "tangent" and as such the inputs can be removed and rely on the default value too, and no space-conversion is necessary.