Adding legacy knurl (#1439)
* Adding legacy knurl

Adding nodedef and nodegraph for Legacy Knurl.
This implements a generic knurl that is a superset, usable standalone, of the knurl used as a Metal pattern.
The original knurl in RapidRT was not usable standalone. This solution is different and has more features, making it a worthwhile extra procedural to add to the library.

A coupe of extra fixes are included
- defaults for a couple of size parameters were incorrectly set to 0,0 after changing them to vector2.
- A converter node from LookdevX was removed. It was unconnected, so no real change, and it might be recreated if that graph is re-exported. I'll try to keep an eye on it.

* Adding diagonal option

The diagonal option was in my original version, but was removed when I added separate X and Y scales, as the pattern was distorting and it was not necessary for matching the RRT version.

But then I realized this was an important feature of Knurl in manufacturing, and spend some time trying to make it work again.

This change adds just a few nodes and the feature is back, even if there is a small approximation in the normals direction when stretched.

* Adding unit test for knurl

Simple unit test