Move uniform updates to object scope in Web Viewer (#2446)
The current handling of uniforms was per-material, which means that when a material is used on multiple objects, they all share uniforms – thus, they were all rendered with the uniforms of the last user of that material.

Any multi-object files were affected. It just happened to work for the default shader ball because there all objects have the same pivot.

As a side effect of this, the previous on-demand updating of uniforms (using `#_updateTransforms`) isn't possible anymore, so I removed the related code.