Additional input units (#1442)
These inputs are unitless in OGS, so they were left unitless in the MaterialX graphs too. But they obviously need to have a unit to keep a consistent size if the scene unit is changed.
Experiments in Inventor seem to suggest all these values are in Inches, but we will check them again once we can do a side-by-side comparison.

Also renamed an input for Knurl, which was not consistent with the rest of the procedurals.