Adding procedural to normal utility node (#1445)
Most of our legacy procedurals have a Color3 output.
This utility node simply takes the R channel, calls heighttonormal, and then normalmap to generate a Vector3 normal in world space, which is the common input for our custom normals in the Protein graphs.

Since this will happen often in the converter, this node simplifies the code,  avoiding creating the graph nodes by code every time.