Fix interpretation of subsurface weight in MDL (#2536)
The subsurface weight should not be applied to the expected color but instead enable/disable the effect in general.
This means the semantic of the weight is meant for composition rather than turning down the volume absorption and scattering itself.

With the performance optimizations for Standard Surface and OpenPBR this issue surfaced since weights are not 1.0 anymore.