Add derivation for OSL field of view
Historically, we've aligned our OSL render tests with GLSL and other languages using an empirically determined field of view, as we didn't yet understand why a default 45-degree field of view didn't produce the expected results in OSL `testrender`.

We now have a better sense of the underlying issue, where OSL `testrender` uses screen-space coordinates in the [-0.5, 0.5] range, while GLSL and other languages use screen-space coordinates in the [-1, 1] range.

This changelist adds a mathematical derivation for the OSL field of view, and replaces the original empirical value with an exact computed value.