Added check for output nodes in functionalGraphs
There was an issue where when you click on an output node in Functional Graphs in the materialX graph editor, it would crash.

The reason for this was that output nodes return a nullptr on getNode() calls, and so were being missed by the check for that
determines if the node was inside a functional graph. I've added this check for output nodes, and now this crash doesn't occur.

Reviewed-by: Niall Lenihan <nlenihan@wetafx.co.nz>

Signed-off-by: sfinnie <sfinnie@wetafx.co.nz>
[nlenihan@wetafx.co.nz: Added detail from internal MR to commit message]
Signed-off-by: Niall Lenihan <nlenihan@wetafx.co.nz>