Improve `Syntax::makeValidName` function (#2598)
This PR improves the logic around how valid names are generated with consideration for reserved words defined in specific Syntax objects.

Currently the shader generator calls `Syntax::makeIdentifier()` to ensure that variable names are legal. This compares the current name against a map of already used identifiers. The reserved words for a syntax are prepopulated in this map.

There are some places, namely [here](https://github.com/AcademySoftwareFoundation/MaterialX/blob/main/source/MaterialXGenShader/Nodes/CompoundNode.cpp#L40) and [here](https://github.com/AcademySoftwareFoundation/MaterialX/blob/main/source/MaterialXGenShader/Nodes/SourceCodeNode.cpp#L70), where we generate names that need to be legal, but do not call `makeIdentifier()` but instead just call `makeValidName()`.  Which got me to thinking that make `makeValidName()` should be doing the reserved word check.  

Moving the check to `makeValidName()` means we do not have to copy all of the reserved words in to the identifier map for each `ShaderGraph` object we create, as well as plugging the potential bug of being able to generate illegal function names.