Align unused parameter warnings across compilers (#2577)
As noted in by @bernardkwok in [#2385](https://github.com/AcademySoftwareFoundation/MaterialX/issues/2385). We have divergent compiler warnings and errors. This makes developing on a platform with more relaxed rules than others more difficult. 

This PR adds the `-Wunused-parameter` compiler flat to non-windows platforms to bring things in line.  As well as fixing the warnings raised in the code not previously compiled on Windows.

We could make this change in the opposite direction, where the Windows compile flags are relaxed. I'm open to that, but don't develop on Windows so can't do that work.  If that is the desired direction I would hope we can post a PR for that immediately, or merge this and then revert this change when someone is able to relax the Windows warnings.  The current difference between platforms feels like the worst situation for us all to be in.