Consolidate `ShaderGenerator::getImplementation` (#2604)
While thinking about #2603 - and how to refactor to avoid storing the node definition name in the `ShaderNode`, I noticed that `getImplementation()` is very similar, and in some cases identical across different shader generators.

This PR consolidates the common implementation to the base `ShaderGenerator` class, and introduces smaller specialization points where necessary.  The `getImplementation()` function is left `virtual` for now, as there may be downstream generators that are overriding it, but in a future breaking change we should probably consider changing that.

The GLSL and MSL implementations were actually identical, so I moved that to the HwShaderGenerator common base class.